Blocks Game in Unity Engine using C# language
A 3D game in Unity Engine involving a block which grows in size when it collects a coin in the grassland. The goal is to collect maximum number of coins by surviving deadly hurdles in the path. Difficulty escalates in every subsequent level and frequently changing colours may prove to be an eye trick.
Moves can be controlled by arrow keys or any suitable keys (WASD)